shared lib added refactoring
