+++ build.gradle
@@ -33,7 +33,6 @@ subprojects {
         implementation "com.fasterxml.jackson.module:jackson-module-kotlin"
 
         implementation "org.springframework.boot:spring-boot-starter-actuator"
-        implementation "org.springframework.boot:spring-boot-starter-aop"
         implementation "org.springframework.boot:spring-boot-starter-web"
 
         testImplementation "org.springframework.boot:spring-boot-starter-test"
