+++ gradle.properties
@@ -9,4 +9,4 @@ app.version.awsJavaSdk=1.12.238
 app.version.tika=2.3.0
 app.version.commonsIO=2.11.0
 app.version.testcontainers=1.17.2
-app.version.mapstruct=1.4.2.Final
+app.version.mapstruct=1.5.2.Final
