+++ library/build.gradle
@@ -1,0 +1,22 @@
+plugins {
+    id "org.springframework.boot"
+    id "io.spring.dependency-management"
+    id "org.jetbrains.kotlin.jvm"
+    id "org.jetbrains.kotlin.kapt"
+    id "org.jetbrains.kotlin.plugin.spring"
+}
+
+dependencies {
+    kapt "org.springframework:spring-context-indexer"
+    kapt "org.springframework.boot:spring-boot-configuration-processor"
+
+    implementation "org.springframework.boot:spring-boot-starter-aop"
+}
+
+bootJar {
+    enabled = false
+}
+
+jar {
+    enabled = true
+}
