+++ library/src/main/kotlin/com/epam/learn/microservices/fundamentals/dto/IdDTO.kt
@@ -1,0 +1,7 @@
+package com.epam.learn.microservices.fundamentals.dto
+
+import java.io.Serializable
+
+data class IdDTO<ID : Serializable>(
+    val id: ID,
+)
