+++ library/src/main/kotlin/com/epam/learn/microservices/fundamentals/dto/IdListDTO.kt
@@ -1,0 +1,7 @@
+package com.epam.learn.microservices.fundamentals.dto
+
+import java.io.Serializable
+
+data class IdListDTO<ID : Serializable>(
+    val ids: Iterable<ID>,
+)
