+++ resource-service/src/main/kotlin/com/epam/learn/microservices/fundamentals/resource/service/config/WebServerConfig.kt
@@ -1,0 +1,15 @@
+package com.epam.learn.microservices.fundamentals.resource.service.config
+
+import org.springframework.context.annotation.Configuration
+import org.springframework.web.bind.annotation.RestController
+import org.springframework.web.method.HandlerTypePredicate
+import org.springframework.web.servlet.config.annotation.PathMatchConfigurer
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer
+
+@Configuration(proxyBeanMethods = false)
+class WebServerConfig : WebMvcConfigurer {
+
+    override fun configurePathMatch(configurer: PathMatchConfigurer) {
+        configurer.addPathPrefix("/api", HandlerTypePredicate.forAnnotation(RestController::class.java))
+    }
+}
