--- resource-service/src/main/kotlin/com/epam/learn/microservices/fundamentals/resource/service/controller/response/IdListResponse.kt
@@ -1,5 +1,0 @@
-package com.epam.learn.microservices.fundamentals.resource.service.controller.response
-
-import java.io.Serializable
-
-data class IdListResponse<ID : Serializable>(val ids: Iterable<ID>)
