+++ song-service/build.gradle
@@ -23,6 +23,7 @@ dependencies {
 
     runtimeOnly "org.postgresql:postgresql"
 
+    implementation project(":library")
     implementation "org.flywaydb:flyway-core"
     implementation "org.springframework.boot:spring-boot-starter-data-jdbc"
     implementation "org.springframework.boot:spring-boot-starter-validation"
